Crate data

Data structures and logic about the DMW2 game data.

This crate deals with the generic DWM2 game data, not a particular play through. Therefore when it refers to “monster name”, it refers to the name of the monster type (e.g. “Slime”), not the name you made up for your monsters.

Modules

breed

Data structures and logic about the DMW2 breed formulae

monster

Data structures and logic about the DMW2 monsters and families

skill

Data structures and logic for skills

xml

A simple XML parser based on quick-xml.

Structs

GameData

All DWM2 game data. This is the entry point of the whole library.